requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
tavily-python>=0.5.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.0.0
httpx>=0.27.0
streamlit>=1.39.0
tiktoken>=0.8.0
numpy>=1.26.0
